Move ScalaJSOutputAnalyzer as private in the 0.6.x Settings.
That way, it is kept as an implementation detail.
